aita-data-analysis/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
pyyaml>=6.0
sqlite3
pathlib 